Add Dynamic Issues and Pull Requests Badges to README
I added Dynamic Issues and Pull Requests Badges to the README to allow people to see how many Issues and Pull Requests there are more clearly and it adds some extra style to the README. This is common on projects like this so people can quickly see how many issues and pull requests there is. It also just looks cool and is really neat to have.